fixed hexany tuning sorting order issue
